+++ plugins/api-docs/src/components/ApiExplorerPage/ApiExplorerPage.test.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright 2020 The Backstage Authors
+ * Copyright 2021 The Backstage Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,185 +14,31 @@
  * limitations under the License.
  */
 
-import { Entity, RELATION_MEMBER_OF } from '@backstage/catalog-model';
-import { ConfigReader } from '@backstage/core-app-api';
-import { TableColumn, TableProps } from '@backstage/core-components';
-import {
-  ConfigApi,
-  configApiRef,
-  storageApiRef,
-} from '@backstage/core-plugin-api';
-import { CatalogTableRow } from '@backstage/plugin-catalog';
-import {
-  CatalogApi,
-  catalogApiRef,
-  DefaultStarredEntitiesApi,
-  entityRouteRef,
-  starredEntitiesApiRef,
-} from '@backstage/plugin-catalog-react';
-import {
-  MockStorageApi,
-  TestApiProvider,
-  wrapInTestApp,
-} from '@backstage/test-utils';
-import DashboardIcon from '@material-ui/icons/Dashboard';
-import { render } from '@testing-library/react';
 import React from 'react';
-import { apiDocsConfigRef } from '../../config';
+import { renderInTestApp } from '@backstage/test-utils';
+import { useOutlet } from 'react-router';
 import { ApiExplorerPage } from './ApiExplorerPage';
 
-describe('ApiCatalogPage', () => {
-  const catalogApi: Partial<CatalogApi> = {
-    getEntities: () =>
-      Promise.resolve({
-        items: [
-          {
-            apiVersion: 'backstage.io/v1alpha1',
-            kind: 'API',
-            metadata: {
-              name: 'Entity1',
-            },
-            spec: { type: 'openapi' },
-          },
-        ] as Entity[],
-      }),
-    getLocationByEntity: () =>
-      Promise.resolve({ id: 'id', type: 'github', target: 'url' }),
-    getEntityByName: async entityName => {
-      return {
-        apiVersion: 'backstage.io/v1alpha1',
-        kind: 'User',
-        metadata: { name: entityName.name },
-        relations: [
-          {
-            type: RELATION_MEMBER_OF,
-            target: { namespace: 'default', kind: 'Group', name: 'tools' },
-          },
-        ],
-      };
-    },
-  };
+jest.mock('react-router', () => ({
+  ...jest.requireActual('react-router'),
+  useOutlet: jest.fn().mockReturnValue('Route Children'),
+}));
 
-  const configApi: ConfigApi = new ConfigReader({
-    organization: {
-      name: 'My Company',
-    },
+jest.mock('./DefaultApiExplorerPage', () => ({
+  DefaultApiExplorerPage: jest.fn().mockReturnValue('DefaultApiExplorerPage'),
+}));
+
+describe('ApiExplorerPage', () => {
+  it('renders provided router element', async () => {
+    const { getByText } = await renderInTestApp(<ApiExplorerPage />);
+
+    expect(getByText('Route Children')).toBeInTheDocument();
   });
 
-  const apiDocsConfig = {
-    getApiDefinitionWidget: () => undefined,
-  };
+  it('renders DefaultApiExplorerPage home when no router children are provided', async () => {
+    (useOutlet as jest.Mock).mockReturnValueOnce(null);
+    const { getByText } = await renderInTestApp(<ApiExplorerPage />);
 
-  const storageApi = MockStorageApi.create();
-
-  const renderWrapped = (children: React.ReactNode) =>
-    render(
-      wrapInTestApp(
-        <TestApiProvider
-          apis={[
-            [catalogApiRef, catalogApi],
-            [configApiRef, configApi],
-            [storageApiRef, storageApi],
-            [
-              starredEntitiesApiRef,
-              new DefaultStarredEntitiesApi({ storageApi }),
-            ],
-            [apiDocsConfigRef, apiDocsConfig],
-          ]}
-        >
-          {children}
-        </TestApiProvider>,
-        {
-          mountedRoutes: {
-            '/catalog/:namespace/:kind/:name': entityRouteRef,
-          },
-        },
-      ),
-    );
-
-  // this test right now causes some red lines in the log output when running tests
-  // related to some theme issues in mui-table
-  // https://github.com/mbrn/material-table/issues/1293
-  it('should render', async () => {
-    const { findByText } = renderWrapped(<ApiExplorerPage />);
-    expect(await findByText(/My Company API Explorer/)).toBeInTheDocument();
-  });
-
-  it('should render the default column of the grid', async () => {
-    const { getAllByRole } = renderWrapped(<ApiExplorerPage />);
-
-    const columnHeader = getAllByRole('button').filter(
-      c => c.tagName === 'SPAN',
-    );
-    const columnHeaderLabels = columnHeader.map(c => c.textContent);
-
-    expect(columnHeaderLabels).toEqual([
-      'Name',
-      'System',
-      'Owner',
-      'Type',
-      'Lifecycle',
-      'Description',
-      'Tags',
-      'Actions',
-    ]);
-  });
-
-  it('should render the custom column passed as prop', async () => {
-    const columns: TableColumn<CatalogTableRow>[] = [
-      { title: 'Foo', field: 'entity.foo' },
-      { title: 'Bar', field: 'entity.bar' },
-      { title: 'Baz', field: 'entity.spec.lifecycle' },
-    ];
-    const { getAllByRole } = renderWrapped(
-      <ApiExplorerPage columns={columns} />,
-    );
-
-    const columnHeader = getAllByRole('button').filter(
-      c => c.tagName === 'SPAN',
-    );
-    const columnHeaderLabels = columnHeader.map(c => c.textContent);
-
-    expect(columnHeaderLabels).toEqual(['Foo', 'Bar', 'Baz', 'Actions']);
-  });
-
-  it('should render the default actions of an item in the grid', async () => {
-    const { findByTitle, findByText } = await renderWrapped(
-      <ApiExplorerPage />,
-    );
-    expect(await findByText(/All \(1\)/)).toBeInTheDocument();
-    expect(await findByTitle(/View/)).toBeInTheDocument();
-    expect(await findByTitle(/View/)).toBeInTheDocument();
-    expect(await findByTitle(/Edit/)).toBeInTheDocument();
-    expect(await findByTitle(/Add to favorites/)).toBeInTheDocument();
-  });
-
-  it('should render the custom actions of an item passed as prop', async () => {
-    const actions: TableProps<CatalogTableRow>['actions'] = [
-      () => {
-        return {
-          icon: () => <DashboardIcon fontSize="small" />,
-          tooltip: 'Foo Action',
-          disabled: false,
-          onClick: () => jest.fn(),
-        };
-      },
-      () => {
-        return {
-          icon: () => <DashboardIcon fontSize="small" />,
-          tooltip: 'Bar Action',
-          disabled: true,
-          onClick: () => jest.fn(),
-        };
-      },
-    ];
-
-    const { findByTitle, findByText } = await renderWrapped(
-      <ApiExplorerPage actions={actions} />,
-    );
-    expect(await findByText(/All \(1\)/)).toBeInTheDocument();
-    expect(await findByTitle(/Foo Action/)).toBeInTheDocument();
-    expect(await findByTitle(/Bar Action/)).toBeInTheDocument();
-    expect((await findByTitle(/Bar Action/)).firstChild).toBeDisabled();
+    expect(getByText('DefaultApiExplorerPage')).toBeInTheDocument();
   });
 });