+++ plugins/kubernetes-backend/src/cluster-locator/ConfigClusterLocator.test.ts
@@ -176,4 +176,76 @@ describe('ConfigClusterLocator', () => {
       },
     ]);
   });
+
+  it('one cluster with dashboardParameters', async () => {
+    const config: Config = new ConfigReader({
+      clusters: [
+        {
+          name: 'cluster1',
+          url: 'http://localhost:8080',
+          authProvider: 'serviceAccount',
+          dashboardApp: 'gke',
+          dashboardParameters: {
+            projectId: 'some-project',
+            region: 'some-region',
+            clusterName: 'cluster1',
+          },
+        },
+      ],
+    });
+
+    const sut = ConfigClusterLocator.fromConfig(config);
+
+    const result = await sut.getClusters();
+
+    expect(result).toStrictEqual([
+      {
+        name: 'cluster1',
+        serviceAccountToken: undefined,
+        url: 'http://localhost:8080',
+        authProvider: 'serviceAccount',
+        skipMetricsLookup: false,
+        skipTLSVerify: false,
+        caData: undefined,
+        dashboardApp: 'gke',
+        dashboardParameters: {
+          projectId: 'some-project',
+          region: 'some-region',
+          clusterName: 'cluster1',
+        },
+      },
+    ]);
+  });
+
+  it('one cluster with dashboardUrl', async () => {
+    const config: Config = new ConfigReader({
+      clusters: [
+        {
+          name: 'cluster1',
+          url: 'http://localhost:8080',
+          authProvider: 'serviceAccount',
+          dashboardApp: 'standard',
+          dashboardUrl: 'http://someurl',
+        },
+      ],
+    });
+
+    const sut = ConfigClusterLocator.fromConfig(config);
+
+    const result = await sut.getClusters();
+
+    expect(result).toStrictEqual([
+      {
+        name: 'cluster1',
+        serviceAccountToken: undefined,
+        url: 'http://localhost:8080',
+        authProvider: 'serviceAccount',
+        skipMetricsLookup: false,
+        skipTLSVerify: false,
+        caData: undefined,
+        dashboardApp: 'standard',
+        dashboardUrl: 'http://someurl',
+      },
+    ]);
+  });
 });